requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
python-dateutil
kaleido
python-pptx
python-docx
requests
scipy

